get_exchange_rate/requirements.txt
python-crontab==2.6.0
requests==2.26.0
xmltodict==0.12.0
psycopg2-binary==2.9.2